2025-1A-T05-CC11-G02-main/2025-1A-T05-CC11-G02-main/app/backend/pyproject.toml
[project]
name = "hybrid-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "ipython>=9.0.2",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "openai>=1.70.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.1",
    "pydantic-ai>=0.0.49",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "uvicorn>=0.34.0",
]
